% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{buildClusterTree}
\alias{buildClusterTree}
\title{Perform hierarchical clustering on reducedDim}
\usage{
buildClusterTree(sce, reduction, labelCol)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object}

\item{reduction}{field of reducedDims(sce) to use}

\item{labelCol}{column in \code{SingleCellExperiment} storing cell type annotations}
}
\description{
Perform hierarchical clustering dimension reduction from single cell expression data
}
