% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumblr.R
\name{clr}
\alias{clr}
\title{Centered log ratio transform}
\usage{
clr(counts, pseudocount = 0.5)
}
\arguments{
\item{counts}{count data with samples as rows and variables are columns}

\item{pseudocount}{added to counts to avoid issues with zeros}
}
\value{
matrix of CLR transformed counts
}
\description{
Compute the centered log ratio (CLR) transform of a count matrix.
}
\details{
The CLR of a vector \code{x} of counts in \code{D} categories is defined as 
\code{clr(x) = log(x) - mean(log(x))}. For details see van den Boogaart and  Tolosana-Delgado (2013).
}
\references{
{
  \insertRef{van2013analyzing}{crumblr}
}
}
\seealso{
\code{compositions::clr}
}
