% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumblr.R
\name{crumblr-package}
\alias{crumblr-package}
\title{crumblr}
\description{
Crumblr enables analysis of count ratio data using precision weighted linear (mixed) models.  It uses an asymptotic normal approximation of the variance following the centered log ration transform (CLR) that is widely used in compositional data analysis.  Crumblr provides a fast, flexible alternative to GLMs and GLMM's while retaining high power and controlling the false positive rate.
}
