% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmn_mle.R
\name{Dirichlet.multinomial}
\alias{Dirichlet.multinomial}
\title{Generation of Dirichlet-Multinomial Random Samples}
\usage{
Dirichlet.multinomial(Nrs, shape)
}
\arguments{
\item{Nrs}{A vector specifying the number of reads or sequence depth for each sample.}

\item{shape}{A vector of Dirichlet parameters for each taxa.}
}
\description{
Random generation of Dirichlet-Multinomial samples.
}
\details{
Adapted from HMP::Dirichlet.multinomial(), since HMP was removed from CRAN
}
\keyword{internal}
