% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotScatterDensity}
\alias{plotScatterDensity}
\title{Scatter plot with 2D density using viridis colors}
\usage{
plotScatterDensity(x, y, size = 1)
}
\arguments{
\item{x}{the x-coordinates of points in the plot}

\item{y}{the y-coordinates of points in the plot}

\item{size}{size of point}
}
\value{
plot from ggplot2
}
\description{
Scatter plot with 2D density using viridis colors
}
\examples{
# simulate data
M = rmvnorm(1000, mu=c(0,0), sigma=diag(1,2))

# create 2D density plot
plotScatterDensity(M[,1], M[,2])

}
