% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst.R
\docType{methods}
\name{vst}
\alias{vst}
\alias{vst,EList-method}
\title{Variance stabilizing transform from precision weights}
\usage{
vst(x)

\S4method{vst}{EList}(x)
}
\arguments{
\item{x}{object storing data to be transformed}
}
\description{
Compute variance stabilizing transform (VST) from precision weights by scaling each observation by their respective weights
}
\details{
A variance stabilizing transform is usually described in terms of a parametric model of the observed data.  Instead, here inverse variance of each observation are stored in \code{x$weight} and the VST divides the observed data by the variances's.
}
