% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IFNCellCounts}
\alias{IFNCellCounts}
\alias{info}
\title{Cell counts following interferon treatment}
\format{
\itemize{    
\item \code{cellCounts} data.frame of cell counts.
\item \code{info} is metadata for each sample
}
}
\usage{
data(IFNCellCounts)
}
\description{
Counts are from single cell RNA-seq data from treated and untreated samples from \insertCite{kang2018multiplexed}{crumblr}.
}
\examples{
data(IFNCellCounts)
}
\references{
{
  \insertRef{kang2018multiplexed}{crumblr}
}
}
\keyword{datasets}
