% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{meanSdPlot}
\alias{meanSdPlot}
\title{Plot row standard deviations versus row means}
\usage{
meanSdPlot(x)
}
\arguments{
\item{x}{data matrix}
}
\value{
plot from ggplot2
}
\description{
Diagnositic plot for variance stabilizing transform
}
\details{
Plot the sd versus rank mean of each row like \code{vsn::meanSdPlot}.  Also show the coefficient of variation of the variances.  A a lower value indicates stronger variance stabilization
}
\examples{
# set probability of each category
prob = runif(300)

# number of samples
n_samples = 1000

# number of counts
nCounts = 300

# simulate counts from multinomial
counts = t(rmultinom(n_samples, size = nCounts, prob = prob))
colnames(counts) = paste0("cat_", 1:length(prob))
rownames(counts) = paste0("sample_", 1:n_samples)

# run crumblr on counts
cobj = crumblr(counts)

# run variance stabilizing transform
df_vst = vst(cobj)

# For each sample, plot rank of mean vs sd
meanSdPlot(t(df_vst)) + ggtitle("crumblr vst")

}
\seealso{
\code{vsn::meanSdPlot}
}
