% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{plotTreeTest}
\alias{plotTreeTest}
\title{Plot tree with results from multivariate testing}
\usage{
plotTreeTest(tree, low = "grey90", mid = "red", high = "darkred")
}
\arguments{
\item{tree}{phylo object storing tree}

\item{low}{low color on gradient}

\item{mid}{mid color on gradient}

\item{high}{high color on gradient}
}
\description{
Plot tree with results from multivariate testing
}
\examples{
library(variancePartition)

# Load cell counts from Kang, et al. (2018)
#  https://doi.org/10.1038/nbt.4042
data(IFNCellCounts)

# Apply crumblr transformation 
cobj = crumblr(cellCounts)

# Use dream workflow to analyze each cell separately
fit = dream(cobj, ~ StimStatus + ind, info)
fit = eBayes(fit)

# Create a hierarchical cluster of cell types
# NOTE: for example only
# Create clustering using prior knowledge 
# or single cell data
hc = hclust(dist(t(cellCounts)))

# Perform multivariate test across the hierarchy
res = treeTest( fit, cobj, hc, coef="StimStatusstim", method="RE2C")

# Plot hierarchy and testing results
# Adjust xlim() until text fits in window 
plotTreeTest(res) + xlim(0, 7)

}
